Moved web/store package to backend/ package for consistency and forward clarity.
