+++ test/store/artifact/fixture.py
@@ -14,8 +14,11 @@ import pytest
 from fastapi.testclient import TestClient
 from sqlalchemy.pool import StaticPool
 
-import mlte.web.store.app_factory as app_factory
+import mlte.backend.app_factory as app_factory
 from mlte.artifact.type import ArtifactType
+from mlte.backend.api.api import api_router
+from mlte.backend.core.config import settings
+from mlte.backend.state import state
 from mlte.store.artifact.factory import create_store
 from mlte.store.artifact.underlying.fs import LocalFileSystemStore
 from mlte.store.artifact.underlying.http import (
@@ -26,9 +29,6 @@ from mlte.store.artifact.underlying.http import (
 from mlte.store.artifact.underlying.memory import InMemoryStore
 from mlte.store.artifact.underlying.rdbs.store import RelationalDBStore
 from mlte.store.base import StoreURI
-from mlte.web.store.api.api import api_router
-from mlte.web.store.core.config import settings
-from mlte.web.store.state import state
 
 _STORE_FIXTURE_NAMES = ["http_store", "memory_store", "fs_store", "rdbs_store"]
